+++ bindings/python/src/lib.rs
@@ -99,7 +99,7 @@ impl StylesheetCache {
     }
 }
 
-/// `CSSInliner(inline_style_tags=True, keep_style_tags=False, keep_link_tags=False, keep_at_rules=False, minify_css=False, base_url=None, load_remote_stylesheets=True, cache=None, extra_css=None, preallocate_node_capacity=32)`
+/// `CSSInliner(inline_style_tags=True, keep_style_tags=False, keep_link_tags=False, keep_at_rules=False, minify_css=False, base_url=None, load_remote_stylesheets=True, cache=None, extra_css=None, preallocate_node_capacity=32, remove_inlined_selectors=False)`
 ///
 /// Customizable CSS inliner.
 #[pyclass]
@@ -108,7 +108,7 @@ struct CSSInliner {
 }
 
 macro_rules! inliner {
-    ($inline_style_tags:expr, $keep_style_tags:expr, $keep_link_tags:expr, $keep_at_rules:expr, $minify_css:expr, $base_url:expr, $load_remote_stylesheets:expr, $cache:expr, $extra_css:expr, $preallocate_node_capacity:expr) => {{
+    ($inline_style_tags:expr, $keep_style_tags:expr, $keep_link_tags:expr, $keep_at_rules:expr, $minify_css:expr, $base_url:expr, $load_remote_stylesheets:expr, $cache:expr, $extra_css:expr, $preallocate_node_capacity:expr, $remove_inlined_selectors:expr) => {{
         let options = rust_inline::InlineOptions {
             inline_style_tags: $inline_style_tags.unwrap_or(true),
             keep_style_tags: $keep_style_tags.unwrap_or(false),
@@ -129,6 +129,7 @@ macro_rules! inliner {
             extra_css: $extra_css.map(Into::into),
             preallocate_node_capacity: $preallocate_node_capacity.unwrap_or(32),
             resolver: std::sync::Arc::new(rust_inline::DefaultStylesheetResolver),
+            remove_inlined_selectors: $remove_inlined_selectors.unwrap_or(false),
         };
         rust_inline::CSSInliner::new(options)
     }};
@@ -138,7 +139,7 @@ macro_rules! inliner {
 impl CSSInliner {
     #[new]
     #[pyo3(
-        signature = (inline_style_tags=true, keep_style_tags=false, keep_link_tags=false, keep_at_rules=false, minify_css=false, base_url=None, load_remote_stylesheets=true, cache=None, extra_css=None, preallocate_node_capacity=32)
+        signature = (inline_style_tags=true, keep_style_tags=false, keep_link_tags=false, keep_at_rules=false, minify_css=false, base_url=None, load_remote_stylesheets=true, cache=None, extra_css=None, preallocate_node_capacity=32, remove_inlined_selectors=false)
     )]
     #[allow(clippy::too_many_arguments)]
     fn new(
@@ -152,6 +153,7 @@ impl CSSInliner {
         cache: Option<StylesheetCache>,
         extra_css: Option<String>,
         preallocate_node_capacity: Option<usize>,
+        remove_inlined_selectors: Option<bool>,
     ) -> PyResult<Self> {
         let inner = inliner!(
             inline_style_tags,
@@ -163,7 +165,8 @@ impl CSSInliner {
             load_remote_stylesheets,
             cache,
             extra_css,
-            preallocate_node_capacity
+            preallocate_node_capacity,
+            remove_inlined_selectors
         );
         Ok(CSSInliner { inner })
     }
@@ -208,12 +211,12 @@ impl CSSInliner {
     }
 }
 
-/// `inline(html, inline_style_tags=True, keep_style_tags=False, keep_link_tags=False, keep_at_rules=False, minify_css=False, base_url=None, load_remote_stylesheets=True, cache=None, extra_css=None, preallocate_node_capacity=32)`
+/// `inline(html, inline_style_tags=True, keep_style_tags=False, keep_link_tags=False, keep_at_rules=False, minify_css=False, base_url=None, load_remote_stylesheets=True, cache=None, extra_css=None, preallocate_node_capacity=32, remove_inlined_selectors=False)`
 ///
 /// Inline CSS in the given HTML document
 #[pyfunction]
 #[pyo3(
-    signature = (html, inline_style_tags=true, keep_style_tags=false, keep_link_tags=false, keep_at_rules=false, minify_css=false, base_url=None, load_remote_stylesheets=true, cache=None, extra_css=None, preallocate_node_capacity=32)
+    signature = (html, inline_style_tags=true, keep_style_tags=false, keep_link_tags=false, keep_at_rules=false, minify_css=false, base_url=None, load_remote_stylesheets=true, cache=None, extra_css=None, preallocate_node_capacity=32, remove_inlined_selectors=false)
 )]
 #[allow(clippy::too_many_arguments)]
 fn inline(
@@ -228,6 +231,7 @@ fn inline(
     cache: Option<StylesheetCache>,
     extra_css: Option<&str>,
     preallocate_node_capacity: Option<usize>,
+    remove_inlined_selectors: Option<bool>,
 ) -> PyResult<String> {
     let inliner = inliner!(
         inline_style_tags,
@@ -239,17 +243,18 @@ fn inline(
         load_remote_stylesheets,
         cache,
         extra_css,
-        preallocate_node_capacity
+        preallocate_node_capacity,
+        remove_inlined_selectors
     );
     Ok(inliner.inline(html).map_err(InlineErrorWrapper)?)
 }
 
-/// `inline_fragment(html, css, inline_style_tags=True, keep_style_tags=False, keep_link_tags=False, keep_at_rules=False, minify_css=False, base_url=None, load_remote_stylesheets=True, cache=None, extra_css=None, preallocate_node_capacity=32)`
+/// `inline_fragment(html, css, inline_style_tags=True, keep_style_tags=False, keep_link_tags=False, keep_at_rules=False, minify_css=False, base_url=None, load_remote_stylesheets=True, cache=None, extra_css=None, preallocate_node_capacity=32, remove_inlined_selectors=False)`
 ///
 /// Inline CSS in the given HTML fragment
 #[pyfunction]
 #[pyo3(
-    signature = (html, css, inline_style_tags=true, keep_style_tags=false, keep_link_tags=false, keep_at_rules=false, minify_css=false, base_url=None, load_remote_stylesheets=true, cache=None, extra_css=None, preallocate_node_capacity=32)
+    signature = (html, css, inline_style_tags=true, keep_style_tags=false, keep_link_tags=false, keep_at_rules=false, minify_css=false, base_url=None, load_remote_stylesheets=true, cache=None, extra_css=None, preallocate_node_capacity=32, remove_inlined_selectors=false)
 )]
 #[allow(clippy::too_many_arguments)]
 fn inline_fragment(
@@ -265,6 +270,7 @@ fn inline_fragment(
     cache: Option<StylesheetCache>,
     extra_css: Option<&str>,
     preallocate_node_capacity: Option<usize>,
+    remove_inlined_selectors: Option<bool>,
 ) -> PyResult<String> {
     let inliner = inliner!(
         inline_style_tags,
@@ -276,19 +282,20 @@ fn inline_fragment(
         load_remote_stylesheets,
         cache,
         extra_css,
-        preallocate_node_capacity
+        preallocate_node_capacity,
+        remove_inlined_selectors
     );
     Ok(inliner
         .inline_fragment(html, css)
         .map_err(InlineErrorWrapper)?)
 }
 
-/// `inline_many(htmls, inline_style_tags=True, keep_style_tags=False, keep_link_tags=False, keep_at_rules=False, minify_css=False, base_url=None, load_remote_stylesheets=True, cache=None, extra_css=None, preallocate_node_capacity=32)`
+/// `inline_many(htmls, inline_style_tags=True, keep_style_tags=False, keep_link_tags=False, keep_at_rules=False, minify_css=False, base_url=None, load_remote_stylesheets=True, cache=None, extra_css=None, preallocate_node_capacity=32, remove_inlined_selectors=False)`
 ///
 /// Inline CSS in multiple HTML documents
 #[pyfunction]
 #[pyo3(
-    signature = (htmls, inline_style_tags=true, keep_style_tags=false, keep_link_tags=false, keep_at_rules=false, minify_css=false, base_url=None, load_remote_stylesheets=true, cache=None, extra_css=None, preallocate_node_capacity=32)
+    signature = (htmls, inline_style_tags=true, keep_style_tags=false, keep_link_tags=false, keep_at_rules=false, minify_css=false, base_url=None, load_remote_stylesheets=true, cache=None, extra_css=None, preallocate_node_capacity=32, remove_inlined_selectors=false)
 )]
 #[allow(clippy::too_many_arguments)]
 fn inline_many(
@@ -303,6 +310,7 @@ fn inline_many(
     cache: Option<StylesheetCache>,
     extra_css: Option<&str>,
     preallocate_node_capacity: Option<usize>,
+    remove_inlined_selectors: Option<bool>,
 ) -> PyResult<Vec<String>> {
     let inliner = inliner!(
         inline_style_tags,
@@ -314,7 +322,8 @@ fn inline_many(
         load_remote_stylesheets,
         cache,
         extra_css,
-        preallocate_node_capacity
+        preallocate_node_capacity,
+        remove_inlined_selectors
     );
     inline_many_impl(&inliner, htmls)
 }
@@ -332,12 +341,12 @@ fn inline_many_impl(
     Ok(output.map_err(InlineErrorWrapper)?)
 }
 
-/// `inline_many_fragments(htmls, css, inline_style_tags=True, keep_style_tags=False, keep_link_tags=False, keep_at_rules=False, minify_css=False, base_url=None, load_remote_stylesheets=True, cache=None, extra_css=None, preallocate_node_capacity=32)`
+/// `inline_many_fragments(htmls, css, inline_style_tags=True, keep_style_tags=False, keep_link_tags=False, keep_at_rules=False, minify_css=False, base_url=None, load_remote_stylesheets=True, cache=None, extra_css=None, preallocate_node_capacity=32, remove_inlined_selectors=False)`
 ///
 /// Inline CSS in multiple HTML fragments
 #[pyfunction]
 #[pyo3(
-    signature = (htmls, css, inline_style_tags=true, keep_style_tags=false, keep_link_tags=false, keep_at_rules=false, minify_css=false, base_url=None, load_remote_stylesheets=true, cache=None, extra_css=None, preallocate_node_capacity=32)
+    signature = (htmls, css, inline_style_tags=true, keep_style_tags=false, keep_link_tags=false, keep_at_rules=false, minify_css=false, base_url=None, load_remote_stylesheets=true, cache=None, extra_css=None, preallocate_node_capacity=32, remove_inlined_selectors=false)
 )]
 #[allow(clippy::too_many_arguments)]
 fn inline_many_fragments(
@@ -353,6 +362,7 @@ fn inline_many_fragments(
     cache: Option<StylesheetCache>,
     extra_css: Option<&str>,
     preallocate_node_capacity: Option<usize>,
+    remove_inlined_selectors: Option<bool>,
 ) -> PyResult<Vec<String>> {
     let inliner = inliner!(
         inline_style_tags,
@@ -364,7 +374,8 @@ fn inline_many_fragments(
         load_remote_stylesheets,
         cache,
         extra_css,
-        preallocate_node_capacity
+        preallocate_node_capacity,
+        remove_inlined_selectors
     );
     inline_many_fragments_impl(&inliner, htmls, css)
 }