+++ bindings/python/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Added
+
+- `remove_inlined_selectors` option to remove selectors that were successfully inlined from `<style>` blocks.
+
 ### Changed
 
 - Update `cssparser` to `0.36`.